Remove old substitution (#178)
* Remove Substitution2

* Remove second order unifier, remove ref to first order unifier

* Remove UnificationUtils2

* scala{fix, fmt}

* Removed FIrstOrderUnifier

* Fix matching tests
