+++ lisa-utils/src/main/scala/lisa/prooflib/Substitution.scala
@@ -3,7 +3,8 @@ package lisa.prooflib
 import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.RunningTheory
 import lisa.kernel.proof.SCProof
-import lisa.kernel.proof.SequentCalculus.*
+import lisa.kernel.proof.SequentCalculus.Sequent
+import lisa.kernel.proof.{SequentCalculus as SC}
 import lisa.prooflib.BasicStepTactic.*
 import lisa.prooflib.ProofTacticLib.*
 import lisa.prooflib.SimpleDeducedSteps
@@ -11,8 +12,8 @@ import lisa.utils.FOLPrinter
 import lisa.utils.KernelHelpers.{_, given}
 import lisa.utils.unification.UnificationUtils
 import lisa.utils.unification.UnificationUtils.getContextFormulaSet
-import lisa.utils.unification.UnificationUtils2
 
+import scala.annotation.nowarn
 import scala.collection.mutable.{Map as MMap}
 
 object Substitution {
@@ -349,5 +350,207 @@ object Substitution {
         }
       }
     }
+  }
+
+  object applySubst extends ProofTactic {
+
+    private def condflat[T](s: Seq[(T, Boolean)]): (Seq[T], Boolean) = (s.map(_._1), s.exists(_._2))
+
+    private def findSubterm2(t: Term, subs: Seq[(VariableLabel, Term)]): (Term, Boolean) = {
+      val eq = subs.find(s => isSameTerm(t, s._2))
+      if (eq.nonEmpty) (eq.get._1(), true)
+      else {
+        val induct = condflat(t.args.map(te => findSubterm2(te, subs)))
+        if (!induct._2) (t, false)
+        else (Term(t.label, induct._1), true)
+
+      }
+
+    }
+
+    private def findSubterm2(f: Formula, subs: Seq[(VariableLabel, Term)]): (Formula, Boolean) = {
+      f match {
+        case PredicateFormula(label, args) =>
+          val induct = condflat(args.map(findSubterm2(_, subs)))
+          if (!induct._2) (f, false)
+          else (PredicateFormula(label, induct._1), true)
+        case ConnectorFormula(label, args) =>
+          val induct = condflat(args.map(findSubterm2(_, subs)))
+          if (!induct._2) (f, false)
+          else (ConnectorFormula(label, induct._1), true)
+        case BinderFormula(label, bound, inner) =>
+          val fv_in_f = subs.flatMap(e => e._2.freeVariables + e._1)
+          if (!fv_in_f.contains(bound)) {
+            val induct = findSubterm2(inner, subs)
+            if (!induct._2) (f, false)
+            else (BinderFormula(label, bound, induct._1), true)
+          } else {
+            val newv = VariableLabel(freshId((f.freeVariables ++ fv_in_f).map(_.id), bound.id))
+            val newInner = substituteVariables(inner, Map(bound -> newv()))
+            val induct = findSubterm2(newInner, subs)
+            if (!induct._2) (f, false)
+            else (BinderFormula(label, newv, induct._1), true)
+          }
+      }
+    }
+
+    private def findSubformula2(f: Formula, subs: Seq[(VariableFormulaLabel, Formula)]): (Formula, Boolean) = {
+      val eq = subs.find(s => isSame(f, s._2))
+      if (eq.nonEmpty) (eq.get._1(), true)
+      else
+        f match {
+          case PredicateFormula(label, args) =>
+            (f, false)
+          case ConnectorFormula(label, args) =>
+            val induct = condflat(args.map(findSubformula2(_, subs)))
+            if (!induct._2) (f, false)
+            else (ConnectorFormula(label, induct._1), true)
+          case BinderFormula(label, bound, inner) =>
+            val fv_in_f = subs.flatMap(_._2.freeVariables)
+            if (!fv_in_f.contains(bound)) {
+              val induct = findSubformula2(inner, subs)
+              if (!induct._2) (f, false)
+              else (BinderFormula(label, bound, induct._1), true)
+            } else {
+              val newv = VariableLabel(freshId((f.freeVariables ++ fv_in_f).map(_.id), bound.id))
+              val newInner = substituteVariables(inner, Map(bound -> newv()))
+              val induct = findSubformula2(newInner, subs)
+              if (!induct._2) (f, false)
+              else (BinderFormula(label, newv, induct._1), true)
+            }
+        }
+    }
+    def findSubterm(t: Term, subs: Seq[(VariableLabel, Term)]): Option[LambdaTermTerm] = {
+      val vars = subs.map(_._1)
+      val r = findSubterm2(t, subs)
+      if (r._2) Some(LambdaTermTerm(vars, r._1))
+      else None
+    }
+
+    def findSubterm(f: Formula, subs: Seq[(VariableLabel, Term)]): Option[LambdaTermFormula] = {
+      val vars = subs.map(_._1)
+      val r = findSubterm2(f, subs)
+      if (r._2) Some(LambdaTermFormula(vars, r._1))
+      else None
+    }
+
+    def findSubformula(f: Formula, subs: Seq[(VariableFormulaLabel, Formula)]): Option[LambdaFormulaFormula] = {
+      val vars = subs.map(_._1)
+      val r = findSubformula2(f, subs)
+      if (r._2) Some(LambdaFormulaFormula(vars, r._1))
+      else None
+    }
+
+    def applyLeftRight(using lib: lisa.prooflib.Library, proof: lib.Proof)(
+        phi: Formula
+    )(premise: proof.Fact)(rightLeft: Boolean = false, toLeft: Boolean = true, toRight: Boolean = true): proof.ProofTacticJudgement = {
+      val originSequent = proof.getSequent(premise)
+      val leftOrigin = ConnectorFormula(And, originSequent.left.toSeq)
+      val rightOrigin = ConnectorFormula(Or, originSequent.right.toSeq)
+
+      if (!toLeft && !toRight) return proof.InvalidProofTactic("applyLeftRight called with no substitution selected (toLeft or toRight).")
+
+      phi match {
+        case PredicateFormula(label, args) if label == equality =>
+          val left = args(0)
+          val right = args(1)
+          val fv_in_phi = (originSequent.left ++ originSequent.right).flatMap(_.freeVariables).map(_.id)
+          val v = VariableLabel(nFreshId(fv_in_phi, 1).head)
+          lazy val isolatedLeft = originSequent.left.filterNot(f => isSame(f, phi)).map(f => (f, findSubterm(f, IndexedSeq(v -> left))))
+          lazy val isolatedRight = originSequent.right.map(f => (f, findSubterm(f, IndexedSeq(v -> left))))
+          if ((!toLeft || isolatedLeft.forall(_._2.isEmpty)) && (!toRight || isolatedRight.forall(_._2.isEmpty)))
+            if (rightLeft)
+              return proof.InvalidProofTactic(s"There is no instance of ${FOLPrinter.prettyTerm(right)} to replace.")
+            else
+              applyLeftRight(equality(right, left))(premise)(true, toLeft, toRight) match {
+                case proof.InvalidProofTactic(m) => return proof.InvalidProofTactic(s"There is no instance of ${FOLPrinter.prettyTerm(left)} to replace.")
+                case v: proof.ValidProofTactic => return v
+              }
+
+          val leftForm = ConnectorFormula(And, isolatedLeft.map((f, ltf) => if (ltf.isEmpty) f else ltf.get.body).toSeq)
+          val rightForm = ConnectorFormula(Or, isolatedRight.map((f, ltf) => if (ltf.isEmpty) f else ltf.get.body).toSeq)
+          val newleft = if (toLeft) isolatedLeft.map((f, ltf) => if (ltf.isEmpty) f else ltf.get(Seq(right))) else originSequent.left
+          val newright = if (toRight) isolatedRight.map((f, ltf) => if (ltf.isEmpty) f else ltf.get(Seq(right))) else originSequent.right
+          val result1: Sequent = (ConnectorFormula(And, newleft.toSeq), phi) |- rightOrigin
+          val result2: Sequent = result1.left |- ConnectorFormula(Or, newright.toSeq)
+
+          var scproof: Seq[SC.SCProofStep] = Seq(SC.Restate(leftOrigin |- rightOrigin, -1))
+          if (toLeft) scproof = scproof :+ SC.LeftSubstEq(result1, scproof.length - 1, List(left -> right), LambdaTermFormula(Seq(v), leftForm))
+          if (toRight) scproof = scproof :+ SC.RightSubstEq(result2, scproof.length - 1, List(left -> right), LambdaTermFormula(Seq(v), rightForm))
+          scproof = scproof :+ SC.Restate(newleft + phi |- newright, scproof.length - 1)
+
+          proof.ValidProofTactic(
+            scproof,
+            Seq(premise)
+          )
+        case ConnectorFormula(label, args) if label == Iff =>
+          val left = args(0)
+          val right = args(1)
+          val fv_in_phi = (originSequent.left ++ originSequent.right).flatMap(_.schematicFormulaLabels).map(_.id)
+          val H = VariableFormulaLabel(nFreshId(fv_in_phi, 1).head)
+          lazy val isolatedLeft = originSequent.left.filterNot(f => isSame(f, phi)).map(f => (f, findSubformula(f, IndexedSeq(H -> left))))
+          lazy val isolatedRight = originSequent.right.map(f => (f, findSubformula(f, IndexedSeq(H -> left))))
+          if ((!toLeft || isolatedLeft.forall(_._2.isEmpty)) && (!toRight || isolatedRight.forall(_._2.isEmpty)))
+            if (rightLeft)
+              return proof.InvalidProofTactic(s"There is no instance of ${FOLPrinter.prettyFormula(right)} to replace.")
+            else
+              applyLeftRight(Iff(right, left))(premise)(true, toLeft, toRight) match {
+                case proof.InvalidProofTactic(m) => return proof.InvalidProofTactic(s"There is no instance of ${FOLPrinter.prettyFormula(left)} to replace.")
+                case v: proof.ValidProofTactic => return v
+              }
+
+          val leftForm = ConnectorFormula(And, isolatedLeft.map((f, ltf) => if (ltf.isEmpty) f else ltf.get.body).toSeq)
+          val rightForm = ConnectorFormula(Or, isolatedRight.map((f, ltf) => if (ltf.isEmpty) f else ltf.get.body).toSeq)
+          val newleft = if (toLeft) isolatedLeft.map((f, ltf) => if (ltf.isEmpty) f else ltf.get(Seq(right))) else originSequent.left
+          val newright = if (toRight) isolatedRight.map((f, ltf) => if (ltf.isEmpty) f else ltf.get(Seq(right))) else originSequent.right
+          val result1: Sequent = (ConnectorFormula(And, newleft.toSeq), phi) |- rightOrigin
+          val result2: Sequent = result1.left |- ConnectorFormula(Or, newright.toSeq)
+
+          var scproof: Seq[SC.SCProofStep] = Seq(SC.Restate(leftOrigin |- rightOrigin, -1))
+          if (toLeft) scproof = scproof :+ SC.LeftSubstIff(result1, scproof.length - 1, List(left -> right), LambdaFormulaFormula(Seq(H), leftForm))
+          if (toRight) scproof = scproof :+ SC.RightSubstIff(result2, scproof.length - 1, List(left -> right), LambdaFormulaFormula(Seq(H), rightForm))
+          scproof = scproof :+ SC.Restate(newleft + phi |- newright, scproof.length - 1)
+
+          proof.ValidProofTactic(
+            scproof,
+            Seq(premise)
+          )
+        case _ => proof.InvalidProofTactic(s"Formula in applySingleSimp need to be of the form a=b or q<=>p and not ${phi.label}")
+      }
+
+    }
+
+    @nowarn("msg=.*the type test for proof.Fact cannot be checked at runtime*")
+    def apply(using
+        lib: lisa.prooflib.Library,
+        proof: lib.Proof,
+        line: sourcecode.Line,
+        file: sourcecode.File
+    )(f: proof.Fact | Formula, rightLeft: Boolean = false, toLeft: Boolean = true, toRight: Boolean = true)(
+        premise: proof.Fact
+    ): proof.ProofTacticJudgement = {
+      f match {
+        case phi: Formula => applyLeftRight(phi)(premise)(rightLeft, toLeft, toRight)
+        case f: proof.Fact =>
+          val seq = proof.getSequent(f)
+          val phi = seq.right.head
+          val sp = new BasicStepTactic.SUBPROOF(using proof)(None)({
+            val x = applyLeftRight(phi)(premise)(rightLeft, toLeft, toRight)
+            proof.library.have(x)
+            proof.library.andThen(SimpleDeducedSteps.Discharge(f))
+          })
+
+          BasicStepTactic.unwrapTactic(sp.judgement.asInstanceOf[proof.ProofTacticJudgement])("Subproof substitution fail.")
+      }
+
+    }
+
+    def toLeft(using lib: lisa.prooflib.Library, proof: lib.Proof, line: sourcecode.Line, file: sourcecode.File)(f: proof.Fact | Formula, rightLeft: Boolean = false)(
+        premise: proof.Fact
+    ): proof.ProofTacticJudgement = apply(f, rightLeft, toLeft = true, toRight = false)(premise)
+
+    def toRight(using lib: lisa.prooflib.Library, proof: lib.Proof, line: sourcecode.Line, file: sourcecode.File)(f: proof.Fact | Formula, rightLeft: Boolean = false)(
+        premise: proof.Fact
+    ): proof.ProofTacticJudgement = apply(f, rightLeft, toLeft = false, toRight = true)(premise)
   }
 }
