+++ src/main/scala/lisa/mathematics/fol/Quantifiers.scala
@@ -3,7 +3,6 @@ package fol
 
 import lisa.automation.kernel.OLPropositionalSolver.Tautology
 import lisa.automation.kernel.SimplePropositionalSolver.*
-import lisa.automation.kernel.SimpleSimplifier.*
 import lisa.prooflib.Substitution
 
 /**
