+++ src/main/scala/lisa/mathematics/settheory/orderings/InclusionOrders.scala
@@ -4,7 +4,6 @@ package orderings
 
 import lisa.automation.kernel.OLPropositionalSolver.Tautology
 import lisa.automation.kernel.SimplePropositionalSolver.*
-import lisa.automation.kernel.SimpleSimplifier.*
 import lisa.automation.settheory.SetTheoryTactics.*
 import lisa.kernel.proof.SequentCalculus as SC
 import lisa.mathematics.fol.Quantifiers.*
